refactor(correctness): split the in-process check executors out of the runner
The rebase onto the relative-import floor put run.ts back over the
large-file bar (the self-guardrail correctly blocked the union). The
import-resolution and structure executors, the defensive failure-parser
wrapper, and the shared label/remedy constants move to
src/analyzers/correctness/pure-checks.ts; run.ts re-exports the constants
so every consumer keeps its import path. No behavior change.
